feat: add namespaceToDevClassName as babelOption
