+++ packages/babel-plugin/src/utils/dev-classname.js
@@ -11,23 +11,8 @@ import type {
   CompiledNamespaces,
   MutableCompiledNamespaces,
 } from '@stylexjs/shared';
-import path from 'path';
 import StateManager from './state-manager';
 
-// TODO: We will need to maintain the full path to the file eventually
-// Perhaps this can be an option that is passed in.
-export function namespaceToDevClassName(
-  namespace: string,
-  varName: null | string,
-  filename: string,
-): string {
-  // Get the basename of the file without the extension
-  const basename = path.basename(filename).split('.')[0];
-
-  // Build up the class name, and sanitize it of disallowed characters
-  const className = `${basename}__${varName ? `${varName}.` : ''}${namespace}`;
-  return className.replace(/[^.a-zA-Z0-9_-]/g, '');
-}
 
 export function injectDevClassNames(
   obj: CompiledNamespaces,
@@ -36,7 +21,7 @@ export function injectDevClassNames(
 ): CompiledNamespaces {
   const result: MutableCompiledNamespaces = {};
   for (const [key, value] of Object.entries(obj)) {
-    const devClassName = namespaceToDevClassName(
+    const devClassName = state.options.namespaceToDevClassName(
       key,
       varName,
       state.filename ?? 'UnknownFile',
@@ -56,7 +41,7 @@ export function convertToTestStyles(
 ): CompiledNamespaces {
   const result: MutableCompiledNamespaces = {};
   for (const [key, _value] of Object.entries(obj)) {
-    const devClassName = namespaceToDevClassName(
+    const devClassName = state.options.namespaceToDevClassName(
       key,
       varName,
       state.filename ?? 'UnknownFile',
